Fix Anthropic schema: vervang z.number().int() door z.number()
Zod v4 emit voor .int() automatisch minimum/maximum (Number.MIN/MAX_SAFE_INTEGER),
en Anthropic structured output staat geen min/max toe op integer types.
Door alle .int() te vervangen door .number() (met text-hint in description
dat het een geheel getal moet zijn) accepteert Anthropic het schema.

Aangepast in: extract-offerte, detect-remove-regions, detect-drawing-box,
detect-price-zones, apply-corrections.

+++ src/app/api/ai/detect-drawing-box/route.ts
@@ -16,10 +16,10 @@ export const maxDuration = 60
 export const dynamic = 'force-dynamic'
 
 const schema = z.object({
-  x: z.number().int(),
-  y: z.number().int(),
-  w: z.number().int(),
-  h: z.number().int(),
+  x: z.number(),
+  y: z.number(),
+  w: z.number(),
+  h: z.number(),
   reason: z.string().optional(),
 })
